+++ app/components/home/FeaturedProjects.tsx
@@ -1,5 +1,6 @@
+import React from "react";
 import { motion } from "framer-motion";
-import ProjectCard from "@/components/shared/ProjectCard";
+import ProjectCard from "@/components/ProjectCard";
 import { Button } from "@/components/ui/button";
 
 interface Project {
